+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/Maintenance/SymbolCreateVM.cs
@@ -8,8 +8,6 @@ namespace TD.WPF.ViewModels.Maintenance
 {
     internal class SymbolCreateVM : AbstractEditVM
     {
-        public MetadataRepository MetadataRepository { get; }
-
         private const string NoBaseCurrencyDisplayText = "No Base Currency";
         private readonly IReadOnlyList<MarketTypeValueItem> _marketTypeValues;
         private readonly ObservableCollection<BaseCurrencyValueItem> _baseCurrencyValues;
@@ -21,9 +19,8 @@ namespace TD.WPF.ViewModels.Maintenance
             var appSettings = AppSettings.GetAppSettings();
             var databasePath = appSettings.DatabasePath;
 
-            MetadataRepository = new MetadataRepository(databasePath);
-            CurrencyRepository = new CurrencyRepository(databasePath, MetadataRepository);
-            ExchangeRepository = new ExchangeRepository(databasePath, MetadataRepository);
+            CurrencyRepository = new CurrencyRepository(databasePath);
+            ExchangeRepository = new ExchangeRepository(databasePath);
             Currencies = new ObservableCollection<Currency>();
             Exchanges = new ObservableCollection<Exchange>();
             _baseCurrencyValues = new ObservableCollection<BaseCurrencyValueItem>